refactor: 컴포넌트 구조 변경 및 mainpage 부분 리팩토링
