+++ wanchcoach/src/app/(route)/kakaosignup/page.tsx
@@ -4,8 +4,8 @@ import Cookies from "js-cookie";
 import styles from "./kakaosignup.module.css";
 import { BasicInputBox, DateInputBox, TwoCheckBox } from "../../_components/component";
 import { ChangeEvent, useEffect, useState } from "react";
-import SignupStepText from "@/app/_components/Component/Signup/SignupStepText";
-import SignupAgree from "@/app/_components/Component/Signup/SignupAgree";
+import SignupStepText from "@/app/_components/Mainpage/Signup/SignupStepText";
+import SignupAgree from "@/app/_components/Mainpage/Signup/SignupAgree";
 import { useRouter } from "next/navigation";
 import { KakaoLoginController, SignupController } from "@/app/util/controller/userController";
 import { FindMyFamilyIdController } from "@/app/util/controller/familyController";
