+++ wanchcoach/src/app/(route)/mainpage/mainpage.module.css
@@ -2,38 +2,3 @@
   overflow: hidden;
   touch-action: none;
 }
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: end;
-  height: 60px;
-  background-color: #f9f9f9;
-}
-.header_box {
-  display: flex;
-  height: 60px;
-  align-items: center;
-  justify-content: space-around;
-  width: 10%;
-}
-
-.bottombar_container {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  height: 70px;
-  width: 100%;
-  position: absolute;
-  bottom: 0;
-  background-color: white;
-  z-index: 1000;
-}
-.bottombar_box {
-  text-align: -webkit-center;
-  padding: 0 2%;
-}
-.bottombar_text {
-  margin-top: 3px;
-  font-size: 12px;
-  font-weight: 700;
-}
