+++ wanchcoach/src/app/(route)/naversignup/page.tsx
@@ -5,8 +5,8 @@ import styles from "./naversignup.module.css";
 import { BasicInputBox, DateInputBox, TwoCheckBox } from "../../_components/component";
 import { ChangeEvent, useEffect, useState } from "react";
 import dayjs, { Dayjs } from "dayjs";
-import SignupStepText from "@/app/_components/Component/Signup/SignupStepText";
-import SignupAgree from "@/app/_components/Component/Signup/SignupAgree";
+import SignupStepText from "@/app/_components/Mainpage/Signup/SignupStepText";
+import SignupAgree from "@/app/_components/Mainpage/Signup/SignupAgree";
 import { useRouter } from "next/navigation";
 import { NaverLoginController, SignupController } from "@/app/util/controller/userController";
 import { FindMyFamilyIdController } from "@/app/util/controller/familyController";
