+++ wanchcoach/src/app/_components/Mainpage/Drug/DrugBox.tsx
