+++ wanchcoach/src/app/_components/Mainpage/Drug/DrugBoxDetail.tsx
