+++ wanchcoach/src/app/_components/Mainpage/Medical/LocationAgree.tsx
