+++ wanchcoach/src/app/_components/Mainpage/Signup/SignupStepText.tsx
