+++ wanchcoach/src/app/util/fetch/fetchWithAuth.tsx
